expose get_syscall_prefix as utils
